[feat]#70 서버통신 완료
